[IMP] api to give list of values
Basically, I created a new Django endpoint because it was too
fucking complicated to create a Post request to the Django App to
create the pdf and then make it available.

The new endpoint gives access to all the data concerning the
nomenclature. It should be easy enough to use it now.

The next step would be to accept GET requests with enough info as
parameters to provide the whole list of codes/dates/prices. And
create the PDF from the frontend - should be easy enough with a
html2pdf for JS.